add runimentary support for an existing blacklist.
if <wwiv_dir>/goodip.txt exists - that contains your whitelist
with 1 ip address per line.

if <wwiv_dir>/badip.txt exists - that contains your blacklist
with 1 ip address per line.

Start of RBL code exists, but no way to configure it yet.

Also if an unhandled exception is thrown, the server will display
a message box with the text of it. Not great, but better than
silently exiting like it doesn now.